+++ src/max_div/_core/metrics/_diversity/__init__.py
@@ -1,4 +1,5 @@
 from ._enum import DiversityContributionFamily, DiversityMetric
+from ._hybrid_metric import DiversityTerm, HybridDiversityMetric
 from ._objective import (
     DiversityObjective,
     DiversityObjectiveHybrid,